// apps/application-dashboard/src/unused/apis/paymentApi.types.ts
export enum PolicyMutateActionType {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
}

export type PolicyStatusType = {
  status: PolicyResultStatus;
  policy_result_id: string;
  policy_result_created_by: string;
  resource_action_metadata: {
    mutate_action: PolicyMutateActionType;
    data: MapAny;
  };
  can_approve: true;
  approval: {
    id: string;
    status: PolicyResultStatus;
  };
};

export type PolicyDetailsType = {
  id: string;
  name: string;
  description: string;
  resource_action_type_id: string;
  organisation_id: string;
  policy_configurations: CreatePolicyConfigPayload;
  created_by: string;
  created_at: string;
  updated_at: string;
  deleted_at: string;
  status: string;
  status_details: PolicyStatusType;
  action_type: PolicyActionType;
};